// frontend/gear-vault-web/src/api/generated/client/models/CustomerSearchResultDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserDto } from './UserDto';
import type { VehicleDto } from './VehicleDto';
export type CustomerSearchResultDto = {
    customer?: UserDto;
    vehicles?: Array<VehicleDto> | null;
};